Refactor BCheck specific classes to their own package.
