+++ src/main/java/ui/view/pane/storefront/bcheck/BCheckPopupMenu.java
@@ -1,4 +1,4 @@
-package ui.view.pane.storefront;
+package ui.view.pane.storefront.bcheck;
 
 import burp.Burp;
